change transaction id type to long

+++ app/src/main/java/com/bountyhunter/kudo/kudoposretail/api/ApiModel.kt
@@ -47,6 +47,6 @@ data class TransactionRequest (
 )
 
 data class Transaction (
-        val id: Int = 5,
+        val id: Long = 5,
         val qty: Int = 13
 )
